ft_memcpy if not architecture
