+++ libs/libft/Makefile
@@ -41,6 +41,7 @@ ft_strlen.c \
 ft_memset.c \
 ft_bzero.c \
 ft_memcpy.c \
+ft_memcpy_avx.c \
 ft_memcpy_x86.c \
 ft_memmove.c \
 ft_strlcpy.c \
