+++ libs/libft/includes/libft.h
@@ -20,6 +20,8 @@
 # include <stdint.h>
 # include <unistd.h>
 # include <stdarg.h>
+# include <emmintrin.h>
+# include <immintrin.h>
 # include <math.h>
 
 int				ft_atoi(const char *nptr);
@@ -37,6 +39,7 @@ void			*memcpy_x86(void *dest __attribute__((unused)),
 void			*ft_memcpy(void *dst,
 					const void *src,
 					const size_t size);
+void			*memcpy_avx(void *dest, const void *src, size_t n);
 void			*ft_memchr(const void *s, int c, size_t n);
 int				ft_memcmp(const void *s1, const void *s2, size_t n);
 void			*ft_memmove(void *dest, const void *src, size_t n);
